Cleanup
- Use InlineScanMap/Set in more cases

- Delete a bunch of types we no longer need.
