+++ frozen-collections-core/src/macros/macro_api.rs
@@ -238,7 +238,7 @@ mod tests {
             quote!({ "1111", "1112", "1113", "1114", "1115", "1116", "1117" }),
         );
 
-        check_impl(":: ScanSet", quote!({ x, "2", "3", }));
+        check_impl(":: InlineScanSet", quote!({ x, "2", "3", }));
         check_impl(":: FzStringSet", quote!({ x, "2", "3", "4" }));
 
         check_impl(":: FzStringSet", quote!({ x, y, z, a, b, c, d }));
@@ -256,7 +256,7 @@ mod tests {
         check_impl(":: InlineScanSet", quote!({ 1, 2, 3, 4, 5, 60000, 70000 }));
         check_impl(":: InlineHashSet", quote!({ 1, 2, 3, 4, 5, 60000, 70000, 80000 }));
 
-        check_impl(":: ScanSet", quote!({ x, 2, 3, 4, 5, 6, 7 }));
+        check_impl(":: InlineScanSet", quote!({ x, 2, 3, 4, 5, 6, 7 }));
         check_impl(":: FzScalarSet", quote!({ x, 2, 3, 4, 5, 6, 7, 8 }));
     }
 
@@ -275,13 +275,13 @@ mod tests {
         check_impl(":: InlineScanSet", quote!({ "1", "2", "3", }));
         check_impl(":: InlineHashSet", quote!({ "1", "2", "3", "4" }));
 
-        check_impl(":: ScanSet", quote!({ Foo(1), Foo(2), Foo(3),}));
+        check_impl(":: InlineScanSet", quote!({ Foo(1), Foo(2), Foo(3),}));
         check_impl(":: FzOrderedSet", quote!({ Foo(1), Foo(2), Foo(3), Foo(4) }));
 
-        check_impl(":: ScanSet", quote!({ x, 2, 3, 4, 5, 6, 7 }));
+        check_impl(":: InlineScanSet", quote!({ x, 2, 3, 4, 5, 6, 7 }));
         check_impl(":: FzScalarSet", quote!({ x, 2, 3, 4, 5, 6, 7, 8 }));
 
-        check_impl(":: ScanSet", quote!({ x, "2", "3", }));
+        check_impl(":: InlineScanSet", quote!({ x, "2", "3", }));
         check_impl(":: FzStringSet", quote!({ x, "2", "3", "4" }));
     }
 
@@ -300,13 +300,13 @@ mod tests {
         check_impl(":: InlineScanSet", quote!({ "1", "2", "3", }));
         check_impl(":: InlineHashSet", quote!({ "1", "2", "3", "4" }));
 
-        check_impl(":: ScanSet", quote!({ Foo(1), Foo(2), Foo(3), }));
+        check_impl(":: InlineScanSet", quote!({ Foo(1), Foo(2), Foo(3), }));
         check_impl(":: FzHashSet", quote!({ Foo(1), Foo(2), Foo(3), Foo(4) }));
 
-        check_impl(":: ScanSet", quote!({ x, 2, 3, 4, 5, 6, 7 }));
+        check_impl(":: InlineScanSet", quote!({ x, 2, 3, 4, 5, 6, 7 }));
         check_impl(":: FzScalarSet", quote!({ x, 2, 3, 4, 5, 6, 7, 8 }));
 
-        check_impl(":: ScanSet", quote!({ x, "2", "3", }));
+        check_impl(":: InlineScanSet", quote!({ x, "2", "3", }));
         check_impl(":: FzStringSet", quote!({ x, "2", "3", "4" }));
     }
