+++ frozen-collections-core/src/maps/mod.rs
@@ -1,6 +1,5 @@
 //! Specialized read-only map types.
 
-pub use binary_search_map::BinarySearchMap;
 pub use dense_scalar_lookup_map::DenseScalarLookupMap;
 pub use eytzinger_search_map::EytzingerSearchMap;
 pub use hash_map::HashMap;
@@ -8,7 +7,6 @@ pub use iterators::*;
 pub use scan_map::ScanMap;
 pub use sparse_scalar_lookup_map::SparseScalarLookupMap;
 
-mod binary_search_map;
 pub(crate) mod decl_macros;
 mod dense_scalar_lookup_map;
 mod eytzinger_search_map;
